FUTR-129: Example querying for existing connections from framework.
