+++ examples/contact-book/src/app/hooks/useAppContext.ts
@@ -1,0 +1,7 @@
+import { useContext } from 'react';
+
+import { AppContext } from '@/app/components/AppContext';
+
+export const useAppContext = () => {
+  return useContext(AppContext);
+};
