+++ examples/contact-book/src/app/page.tsx
@@ -1,20 +1,27 @@
-import { NextUIProvider } from '@nextui-org/react';
+import { NextUIProvider, Button } from '@nextui-org/react';
 
+import { getSession } from '@/app/actions/session';
 import { Contacts } from '@/app/components/Contacts';
 import { Integrations } from '@/app/components/Integrations';
-
-// import arkwConfig from '@/arkw.config';
+import { SignIn } from '@/app/components/SignIn';
 
 export default async function Home() {
+  const session = await getSession();
+
   return (
     <main className={'h-full'}>
       <NextUIProvider>
-        <div className={'flex flex-col items-center p-4'}>
-          <Integrations />
-          <div>
-            <Contacts />
+        {!session ? (
+          <SignIn />
+        ) : (
+          <div className={'flex flex-col items-center p-4'}>
+            {/*@ts-ignore*/}
+            <Integrations />
+            <div>
+              <Contacts />
+            </div>
           </div>
-        </div>
+        )}
       </NextUIProvider>
     </main>
   );
