--- examples/contact-book/src/arkw.config.ts
@@ -1,121 +1,0 @@
-import { Config, createFramework, extractSchemaOptions } from '@arkw/core';
-import { GoogleIntegration } from '@arkw/google';
-import { MailchimpIntegration } from '@arkw/mailchimp';
-import { z } from 'zod';
-
-export const dbUrl = process.env.DB_URL ?? 'file://contact-book.db';
-export const redirectHost = process.env.APP_URL;
-
-if (!dbUrl || !redirectHost) {
-  throw new Error('Missing required environment variables');
-}
-
-export const REDIRECT_URI = new URL('/api/arkw/connect/callback', redirectHost).toString();
-
-const RECORD_TYPE = { contact: 'contact' } as const;
-
-const BASE_RECORD_SCHEMA = z.object({
-  recordType: z.enum([RECORD_TYPE.contact]),
-});
-
-const RECORD_SCHEMA = z.discriminatedUnion('recordType', [
-  z.object({
-    recordType: z.literal(RECORD_TYPE.contact),
-    data: z.object({
-      firstName: z.string().trim().min(1, 'Required'),
-      lastName: z.string().trim().min(1, 'Required'),
-      email: z.string().email(),
-      birthday: z.date().optional(),
-    }),
-  }),
-]);
-
-export const config: Config = {
-  name: 'contact-book',
-  systemHostURL: process.env.APP_URL!,
-  routeRegistrationPath: '/api/arkw',
-  blueprintDirPath: '/src/blueprints',
-  db: {
-    provider: 'sqlite',
-    uri: dbUrl,
-  },
-  integrations: [
-    new GoogleIntegration({
-      config: {
-        CLIENT_ID: process.env.GOOGLE_CLIENT_ID!,
-        CLIENT_SECRET: process.env.GOOGLE_CLIENT_SECRET!,
-        REDIRECT_URI,
-        TOPIC: process.env.GOOGLE_MAIL_TOPIC!,
-      },
-    }),
-    new MailchimpIntegration({
-      config: {
-        CLIENT_ID: process.env.MAILCHIMP_CLIENT_ID!,
-        CLIENT_SECRET: process.env.MAILCHIMP_CLIENT_SECRET!,
-        REDIRECT_URI,
-      },
-    }),
-  ],
-  systemEvents: [
-    {
-      key: 'record_created',
-      schema: BASE_RECORD_SCHEMA,
-      triggerProperties: {
-        type: 'RECORD_CREATED',
-        label: 'Record Created',
-        icon: {
-          alt: 'Record Created',
-          icon: 'record-created',
-        },
-        description: 'Triggered when a record is created',
-        schema: BASE_RECORD_SCHEMA,
-        async getSchemaOptions() {
-          const options = extractSchemaOptions({ schema: BASE_RECORD_SCHEMA });
-          return options;
-        },
-        outputSchema: RECORD_SCHEMA,
-      },
-    },
-    {
-      key: 'record_updated',
-      schema: BASE_RECORD_SCHEMA,
-      triggerProperties: {
-        type: 'RECORD_UPDATED',
-        label: 'Record Updated',
-        icon: {
-          alt: 'Record Updated',
-          icon: 'edit',
-        },
-        description: 'Triggered when a record is updated',
-        schema: BASE_RECORD_SCHEMA,
-        async getSchemaOptions() {
-          const options = extractSchemaOptions({ schema: BASE_RECORD_SCHEMA });
-          return options;
-        },
-        outputSchema: RECORD_SCHEMA,
-      },
-    },
-    {
-      key: 'record_deleted',
-      schema: BASE_RECORD_SCHEMA,
-      triggerProperties: {
-        type: 'RECORD_DELETED',
-        label: 'Record Deleted',
-        icon: {
-          alt: 'Record Deleted',
-          icon: 'trash',
-        },
-        description: 'Triggered when a record is deleted',
-        schema: BASE_RECORD_SCHEMA,
-        async getSchemaOptions() {
-          const options = extractSchemaOptions({ schema: BASE_RECORD_SCHEMA });
-          return options;
-        },
-        outputSchema: RECORD_SCHEMA,
-      },
-    },
-  ],
-  systemActions: [],
-};
-
-export default createFramework(config);
